refactor: move the main page to a separate module
